Include package.json when user specifies inputs
We want to include each workspace's package.json as an input when
the user has configured a smaller set of inputs, because the package.json
defines the actual task turbo will execute for that workspace (via scripts).
If that task definition changes, we want to consider the cache invalid.

We only apply this inclusion when inputs are defined, because if inputs
are _not_ defined, _all_ git-tracked files are included in the hash,
'which will already include the package.json file.